% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbl_extr_plate}
\alias{dbl_extr_plate}
\title{Tidy Table with 2 different Extractant IDs}
\format{
A tibble similar to \code{tidy_plates}, with one row per well, and the following columns:

\describe{
\item{row,column,well_id,unique_well_id,dataset,plate_id}{allowing the unique
identification of each well of the 96-well plate}
\item{map}{identifier of samples, Standard curve or extractant (“extr_1" or "extr_2")}
\item{abs}{raw absorbance read of each well}
\item{extr_id}{extractant identifier corresponding to each well. 3 possible
values: "extr_1", "extr_2" or "none" (for empty wells or standard curves)}
}
}
\usage{
dbl_extr_plate
}
\description{
This is a modified version of \code{tidy_plates} where some wells have been artificially
attributed to a second extractant. It serves as example to test several steps
of the vignette \code{blank-correction} and functions therein. It has one more column
than \code{tidy_plates} that stores the information of which well needs to be
corrected with which extractant
}
\keyword{datasets}
