% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_std_curve.R
\name{density_lm_param}
\alias{density_lm_param}
\title{Plot Distribution of Model QC Parameters (p-val, R2 and adjusted R2)}
\usage{
density_lm_param(
  lm_data,
  p_or_r = "p",
  threshold = 0.05,
  facetting_std_sp = TRUE,
  color_std_sp = TRUE
)
}
\arguments{
\item{lm_data}{lm_data in the shape as generated by \code{lm_std_curve()},
possibly as a subset to observe only suspicious curves. Must contain a
column named \code{std_sp} (add it with \code{dplyr::left_join()} if necessary)}

\item{p_or_r}{Defines which parameter to plot Accepted values are \code{"p"}
(default), \code{"adjR2"}and \code{"R2"}.}

\item{threshold}{Defaults at 0.05 (as a significance threshold of the p-value).
Accepts any value. For R2 or adjusted R2, 0.95 or 0.98 might be good options.
This is a graphical parameter only (defines the position of the vertical line)}

\item{facetting_std_sp}{Logical, defaults as \code{TRUE}. Whether to use faceting
according to standard species}

\item{color_std_sp}{Logical, defaults as \code{TRUE}. Whether to group according to
standard species (and overplot with several colours)}
}
\value{
A ggplot object
}
\description{
Plot Distribution of Model QC Parameters (p-val, R2 and adjusted R2)
}
\examples{
lm_data <- lm_output$lm_data # no column called "std_sp"
# get standard species from std_corrected
(std_sp <- std_corrected |> dplyr::select(plate_id, std_sp) |> unique())
# Artificially changing a std_sp for the sake of facetting
std_sp$std_sp[3:4] <- rep("NO2")
lm_data_full <- lm_data |> dplyr::left_join(std_sp, by = dplyr::join_by(plate_id))
density_lm_param(lm_data_full, facetting_std_sp = FALSE)
density_lm_param(
  lm_data_full,
  p_or_r = "adjR2", threshold = 0.95,
  color_std_sp = FALSE)
}
