% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve.R
\name{extract_curve}
\alias{extract_curve}
\title{Get concentration of standard curve from metadata}
\usage{
extract_curve(
  metadata,
  pipetting_direction = "top_down",
  dataset_col = "dataset",
  plate_id_col = "plate_id",
  std_conc_col = "std_conc",
  row_col = "row"
)
}
\arguments{
\item{metadata}{A tibble following a similar structure as \code{\link{metadata}}, see documentation of `metadata for more details}

\item{pipetting_direction}{Can only be "top_down" (default) or "bottom_up".
A top_down pipetting means that the curve was pipetted vertically (in a single column of the 96-well plate),
with the smallest value (blank) in row A and the highest value in row H.
Conversely, bottom_up pipetting would have the blank in row H and the most concentrated solution in row A}

\item{dataset_col}{Name of the column identifying the dataset. Defaults
to \code{"dataset"}.}

\item{plate_id_col}{Name of the column identifying physical plates.
Defaults to \code{"plate_id"}.}

\item{std_conc_col}{Name of the column containing the standard curve
concentrations (dash-separated, one value per row A-H). Defaults
to \code{"std_conc"}.}

\item{row_col}{Name of the output column identifying plate row (A-H).
Defaults to \code{"row"}.}
}
\value{
A tibble with 3 columns: \code{plate_id}, \code{row} (corresponding to plate-row,
from A to H) and \code{std_conc} containing the concentrations as given in the \code{std_conc} column of \code{metadata}.
}
\description{
Get concentration of standard curve from metadata
}
\examples{
metadata
extract_curve(metadata)
}
\seealso{
\link{metadata}
}
