% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_blanks.R
\name{extractant_average}
\alias{extractant_average}
\title{Computing the per-plate average for raw absorbance of the extractant (blank for samples)}
\usage{
extractant_average(
  data = NULL,
  extractant_data = NULL,
  extr_def = "extr",
  dataset_col = "dataset",
  plate_id_col = "plate_id",
  map_col = "map",
  value_col = "abs"
)
}
\arguments{
\item{data}{A tibble like \code{tidy_plates}}

\item{extractant_data}{Alternatively, A tibble containing only extractant data
Defaults to NULL, where it would be computed from \code{data} using \code{extract_extractant(data)}}

\item{extr_def}{A string that characterizes wells containing the extractant
in the mapping (\code{map}column) of the plate. Defaults to "extr". Can be a vector
containing several values (see examples)}

\item{dataset_col}{Name of the column identifying the dataset. Defaults
to \code{"dataset"}.}

\item{plate_id_col}{Name of the column identifying physical plates.
Defaults to \code{"plate_id"}.}

\item{map_col}{Name of the column containing well mapping/type
information. Defaults to \code{"map"}.}

\item{value_col}{Name of the numeric absorbance column. Defaults to
\code{"abs"}.}
}
\value{
A tibble with one row per plate, contianing the average, standard
deviation and coefficient of variation of raw absorbance of the extractant
}
\description{
Computing the per-plate average for raw absorbance of the extractant (blank for samples)
}
\examples{
data = tidy_plates
(blank_avg <- extractant_average(data, extr_def = "extr"))

# artificially construct a tibble with 2 extractants and an additional column for extractant id
tidy_2_extr <- tidy_plates |>
  dplyr::mutate(
    map = dplyr::case_when(
      column == 8 ~ "extr_1",
      column == 4 ~ "extr_2",
      .default = map))
multiple_extractant_id # check it out
# joining with multiple_extractant_id
(dbl_extr_plate <- tidy_2_extr |> dplyr::left_join(multiple_extractant_id))
(blank_avg <- extractant_average(dbl_extr_plate, extr_def = c("extr_1", "extr_2")))
}
