% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{failed_wells_example}
\alias{failed_wells_example}
\title{Example of failed well table}
\format{
A tibble 2 rows and 3 columns (dataset, plate_id and well_id), allowing unique identification of failed wells
\describe{
\item{dataset,plate_id}{must be the same as in your plate data}
\item{well_id}{standard well identifiers, concatenation of plate rows (letters from A to H) and plate columns (numbers from 1 to 12)}
}
}
\usage{
failed_wells_example
}
\description{
In experiments with 96-well plates, especially with manual pipetting,
it is not rare to have some mishaps where we know that a certain well is to be excluded.
A template for such data can be obtained using \code{\link[=failed_wells_template]{failed_wells_template()}},
exported (e.g., with \code{\link[readr:write_delim]{readr::write_csv()}}) filled by users and imported (e.g., with \code{\link[readr:read_delim]{readr::read_csv()}}).
}
\keyword{datasets}
