% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failed_wells_template.R
\name{failed_wells_template}
\alias{failed_wells_template}
\title{Generate a template to record "failed wells"}
\usage{
failed_wells_template(nrow = 30)
}
\arguments{
\item{nrow}{The number of rows that the user want to have. Defaults at \code{nrow = 30}.
Note that if data is encoded in an Excel or equivalent file, the number of rows is not really important and can be increased indefinitely}
}
\value{
A simple tibble with \code{nrow} rows and 3 columns: dataset, plate_id and well_id.
}
\description{
In experiments with 96-well plates, especially with manual pipetting,
it is not rare to have some mishaps where we know that a certain well is to be excluded.
The template generated here, once filled by the user with data allowing the identification of failed wells,
will allow the automated exclusion of those wells from data analysis.
The template can then easily be exported, for example with \code{\link[readr:write_delim]{readr::write_csv()}} or \code{\link[readr:write_delim]{readr::write_excel_csv()}}.
}
\examples{
failed_wells_template()
}
