% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verticalize_tidy_join.R
\name{join_abs_map}
\alias{join_abs_map}
\title{Merging 2 vertical plates into one}
\usage{
join_abs_map(
  abs_tibble,
  map_tibble,
  dataset = "",
  abs_map = c("abs-", "map-"),
  coerce_numeric = c(FALSE, FALSE)
)
}
\arguments{
\item{abs_tibble, map_tibble}{The first and second tibble (will appear on the left and right, respectively). Must have the same structure as \code{tibble_example}.}

\item{dataset}{An optional string to be added as a prefix to all column names (from both tibbles), with the exception of the first 2 columns describing well id ("row" and "column"). It is originally meant to record the name of the dataset for later uses.}

\item{abs_map}{A string vector to add additional prefixes. The default value is set to c("abs", "map"), so that the "abs" data (corresponding to argument \code{abs_tibble}) will receive the first prefix, and the "map" data (corresponding to argument \code{map_tibble}) will receive the second prefix. Set this to c("", "") to prevent prefix addition.}

\item{coerce_numeric}{A logical vector to decide whether the function \code{verticalize_plates()}, called separately for \code{abs_tibble} and \code{map_tibble}, should coerce data to become numeric or not. The default value is set to \code{c(FALSE, FALSE)}, so that eventually all data can be pivotted in a single column (see later steps in the pipeline).}
}
\value{
A unique verticalized table containing the data from both data sets.
}
\description{
The function \code{join_abs_map()} was thought to merge absorbance data with their mapping counterparts,
coming from 2 separate import occurrences, into a single, vertical tibble.
It takes profit of the \code{dplyr::left_join()} function, connected to our \code{verticalize_plates()} function,
so that it provides a 2-in-1 feature of verticalizing plates while joining them.
}
\details{
The first purpose of this function is to join an absorbance tibble and a mapping tibble, which is how the default setup is organized.
Still, it offers enough flexibility in its parameters to be adapted to the joining of any 2 tibbles, so long as they fit the proper \code{tibble_example}-like structure.
}
\examples{
skanit_csv <- system.file("extdata", "skanit.csv", package = "plate2N")
skanit_tibbles <- skanit_to_tibble(skanit_csv)
join_abs_map(skanit_tibbles$abs_tibble, skanit_tibbles$map_tibble)
}
\seealso{
\code{\link[=verticalize_plates]{verticalize_plates()}}
}
