% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_output}
\alias{lm_output}
\title{Example of Outputs from the Linear Model of the Standard Curves}
\format{
List containing 3 elements, computed with \code{lm_std_curve()}, \code{suspicious_lm()}
and \code{plot_list_lm()}.

\describe{
\item{lm_data}{Tibble, 1 row per curve, serves as summary for the linear model}
\item{suspicious_lm}{Subset of \code{lm_data}, keeping only "problematic" curves,
i.e., either with p-value > 0.5, or with non-normality or
heteroscedasticity of residuals}
\item{plot_list_suspicious}{A list of ggplots displaying those problematic curves
(only one plot in this data set)}
}
}
\usage{
lm_output
}
\description{
Example of Outputs from the Linear Model of the Standard Curves
}
\keyword{datasets}
