% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{map_1_plate}
\alias{map_1_plate}
\title{Mapping a Vector with Sample ids into a 96-well Plate Format - Single Plate}
\usage{
map_1_plate(
  plate_id,
  samples,
  std_def = "Std",
  column_curves = c(1, 12),
  blank_def = "extr",
  empty_def = "empty",
  column_empty = c(),
  column_blank = 8,
  n_wells_samples = 4
)
}
\arguments{
\item{plate_id}{The unique identifier of the plate}

\item{samples}{A vector of strings containing sample names. Samples will be
displayed vertically in the plate (see examples)}

\item{std_def}{A single string to define wells containing the serial dilutions
of the standard curve. Defaults to "Std"}

\item{column_curves}{Which columns to map the standard curves to. Can be a
unique string or a vector. Defaults to \code{c(1,12)}.}

\item{blank_def}{A single string to define wells containing the blank. Defaults
to \code{extr} (for "extractant")}

\item{empty_def}{A single string to define empty wells}

\item{column_empty}{Optional, which columns to map as empty}

\item{column_blank}{Which column(s) to map to the blank. Defaults to \code{8}.}

\item{n_wells_samples}{How many replicates to attribute to each sample.
Defaults to \code{4}.}
}
\value{
A tibble of a single 96-well plate, in plate format, with the plate
id in the top left corner
}
\description{
Mapping a Vector with Sample ids into a 96-well Plate Format - Single Plate
}
\details{
This mapping function always attributes complete columns to standard curves and
blanks. Additional empty columns may be given, optionally.

If the list of samples is too short to fill all the wells, remaining empty
wells will be mapped using the character string defined under \code{empty_def}.
}
\examples{
(samples <- sample(LETTERS, size = 15))
map_1_plate(plate_id = "test_plate", samples = samples)
}
