% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{map_plates}
\alias{map_plates}
\title{Mapping a Vector with Sample ids into 96-well Plate Format - Several Plates}
\usage{
map_plates(
  samples,
  n_samples_per_plate = 18,
  plate_ids = NULL,
  std_def = "Std",
  column_curves = c(1, 12),
  blank_def = "extr",
  rename_na = "empty",
  empty_def = "empty",
  column_empty = c(),
  column_blank = 8,
  n_wells_samples = 4
)
}
\arguments{
\item{samples}{A vector of strings containing sample names. Samples will be
divided into plates, and displayed vertically (see examples)}

\item{n_samples_per_plate}{How many samples should be fitted on a single plate.
Defaults to 18, which fits the default settings of \code{map_1_plate()}.}

\item{plate_ids}{Optional, a vector of plate_ids to attribute to the plates.
Must be the fitting length. By default, plate ids will be \code{plate_1}, \code{plate_2}, etc.}

\item{std_def, column_curves, blank_def, rename_na, empty_def, column_empty, column_blank, n_wells_samples}{These
parameters are given to a call to \code{map_1_plate()} and follow its logic.
See also \code{?map_1_plate()}.}
}
\value{
A tibble with the plate mapping, in a format similar to \code{tibble_example}.
This format also fits the import pipeline. See also the vignette \code{import-tidy}.
}
\description{
Mapping a Vector with Sample ids into 96-well Plate Format - Several Plates
}
\examples{
# take random fruit names as sample names
samples <- stringr::fruit[1:70] ; samples[1:5]
map_plates(samples, 14) |> print(n = 20)
}
\seealso{
\code{\link[=map_1_plate]{map_1_plate()}}
}
