% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molar_masses}
\alias{molar_masses}
\title{Record of molar masses for studied compounds}
\format{
A vector of named numerics containing relevant molar masses.

\describe{
\item{N}{molar mass of elemental nitrogen in g/mol (not N2)}
\item{NO3}{molar mass of nitrate in g/mol}
\item{NO2}{molar mass of nitrite in g/mol}
\item{NH4}{molar mass of ammonium in g/mol}
}
}
\usage{
molar_masses
}
\description{
Originally, it contains molar masses of N, NO3-, NO2-, NH4+, but it can easily
be appended with additional molar masses for other relevant applications
}
\keyword{datasets}
