% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_std_curve.R
\name{plot_list_lm}
\alias{plot_list_lm}
\title{Compute a list of plots, one per standard curve}
\usage{
plot_list_lm(
  lm_data,
  std_data,
  model = "linear",
  conc_col = "std_conc",
  value_col = "abs_corrected",
  curve_id_col = "unique_curve_id",
  dataset_col = "dataset",
  unit_col = "std_unit"
)
}
\arguments{
\item{lm_data}{lm_data in the shape as generated by \code{lm_std_curve()},
possibly as a subset to observe only suspicious curves}

\item{std_data}{tibble containing only standard data, as in \code{std_corrected}.
Absorbance values should already by blank-corrected}

\item{model}{Which model to use. Accepts either \code{linear} (default) or \code{poly}
for polynomial model.}

\item{conc_col}{Name of the column containing concentration. Defaults
to \code{"std_conc"}.}

\item{value_col}{Name of the column containing (blank-corrected)
absorbance. Defaults to \code{"abs_corrected"}.}

\item{curve_id_col}{Name of the column identifying which curve a row
belongs to. Defaults to \code{"unique_curve_id"}.}

\item{dataset_col}{Name of the column identifying the dataset.
Defaults to \code{"dataset"}.}

\item{unit_col}{Name of the column containing the concentration unit,
shown in the x-axis label. Defaults to \code{"std_unit"}.}
}
\value{
A list of plots, one per curve, annotated with the issue:
- non-significance of the linear modal (p-value > 0.05)
- non-normality of residuals (shapiro test with threshold of p = 0.05)
- heteroscedasticity of residuals (Breusch-Pagan test, threshold of p = 0.05)
- if \code{model = "poly"}: non-significance of the \code{a} and \code{b} coefficients of the
regression equation \code{y = ax^2 + bx} (p-value > 0.05)
}
\description{
Compute a list of plots, one per standard curve
}
\examples{
lm_data <- lm_std_curve(std_corrected |> dplyr::group_by(plate_id, column))
plot_list <- plot_list_lm(
    lm_data,
    std_data = std_corrected)
plot_list[[1]] ; plot_list[[2]]
}
