% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plates.R
\name{read_tecan}
\alias{read_tecan}
\title{Import single plate data from a single Tecan-generated file}
\usage{
read_tecan(file, extension = ".xlsx")
}
\arguments{
\item{file}{(path to) file to import from}

\item{extension}{".xlsx" is currently the only option, but this might change based on user needs}
}
\value{
A tibble with a single plate
}
\description{
Import single plate data from a single Tecan-generated file
}
\examples{
file <- system.file("extdata", "tecan_example/tecan1.xlsx", package = "plate2N")
read_tecan(file)
}
