% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_to_conc.R
\name{reg_join_abs}
\alias{reg_join_abs}
\title{Merge Data from Regression Equation on Standard Curve with Absorbance Data (samples)}
\usage{
reg_join_abs(
  lm_table,
  abs_data,
  target_sp = "N",
  plate_id_col = "plate_id",
  dataset_col = "dataset",
  map_col = "map",
  well_id_col = "well_id",
  value_col = "abs_corrected",
  std_sp_col = "std_sp",
  unit_col = "std_unit"
)
}
\arguments{
\item{lm_table}{A tibble with linear model data (as generated by \code{lm_std_curve()}).
Must contain columns \code{plate_id}, \code{slope}, \code{adj_r_squared}, \code{lm_p} (linear
model) or \code{plate_id}, \code{poly_a}, \code{poly_a_p}, \code{poly_b}, \code{poly_b_p},
\code{r_squared}, \code{adj_r_squared}, \code{lm_p} (polynomial model)}

\item{abs_data}{Blank-corrected absorbance data, in a similar format to tidy_plates,
but containing only sample wells, no empty wells, standard curve wells or extractant wells}

\item{target_sp}{Defaults to "N", the target compound to convert to}

\item{plate_id_col}{Name of the column identifying physical plates.
Defaults to \code{"plate_id"}.}

\item{dataset_col}{Name of the column identifying the dataset. Defaults
to \code{"dataset"}.}

\item{map_col}{Name of the column containing well mapping/type
information. Defaults to \code{"map"}.}

\item{well_id_col}{Name of the column identifying wells. Defaults to
\code{"well_id"}.}

\item{value_col}{Name of the column containing (blank-corrected)
absorbance. Defaults to \code{"abs_corrected"}.}

\item{std_sp_col}{Name of the column identifying the standard species.
Defaults to \code{"std_sp"}.}

\item{unit_col}{Name of the column containing the concentration unit.
Defaults to \code{"std_unit"}.}
}
\value{
A tibble joining linear model data with absorbance data
}
\description{
Merge Data from Regression Equation on Standard Curve with Absorbance Data (samples)
}
\examples{
lm_table <- lm_output$lm_data
sample_data <- sample_corrected
reg_join_abs(lm_table, sample_data, target_sp = "N")
}
