% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_wells.R
\name{remove_wells}
\alias{remove_wells}
\title{Clean a tidy table from undesired wells}
\usage{
remove_wells(table_to_clean, well_table)
}
\arguments{
\item{table_to_clean, well_table}{Tibbles with the same structure, respectively, as \code{\link{tidy_table}} and \code{\link{failed_wells_example}}.
Column (dataset, plate_id, well_id) must have strictly the same name, but the order of columns does not matter.
Rows of \code{well_table} uniquely identify wells to be removed}
}
\value{
A cleaned version of the input tidy_table, that should have less rows than the input table (shortened by the number of rows of \code{well_table})
}
\description{
This function relies on a very simple use of the \code{dplyr::anti_join()} function with settings that are particular to our data structure.
However, its iterative use in analysis justifies the definition of an ad-hoc function.
}
\examples{
print(tidy_table, n = 10); failed_wells_example
tidy_table |> remove_wells(failed_wells_example) |> print(n = 10)
}
