% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_corrected}
\alias{sample_corrected}
\title{Example of Blank-corrected Absorbance Data (samples)}
\format{
Tibble in format as generated by \code{blank_correct_abs()}

\describe{
\item{abs_corrected}{blank-corrected absorbance}
\item{other columns}{Same columns as \code{tidy_plates}, + extractant average}
}
}
\usage{
sample_corrected
}
\description{
Example of Blank-corrected Absorbance Data (samples)
}
\keyword{datasets}
