% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve.R
\name{std_blank_average}
\alias{std_blank_average}
\title{Computing per-plate average for raw absorbance of the standard blank}
\usage{
std_blank_average(
  std_blank_data,
  dataset_col = "dataset",
  plate_id_col = "plate_id",
  value_col = "abs"
)
}
\arguments{
\item{std_blank_data}{A tibble containing only rows that received standard blanks,
i.e., row A (top_down_pipetting) or H (bottom_up_pipetting). It is normally
generated by \code{extract_std_blank()} from which one element can be chosen and,
if necessary, modified, based on the examination of "untrusted wells".
Thus, use as input (a modified version of) \code{extract_std_blank(data)$all} or
\code{extract_std_blank(data)$trusted}}

\item{dataset_col}{Name of the column identifying the dataset. Defaults
to \code{"dataset"}.}

\item{plate_id_col}{Name of the column identifying physical plates.
Defaults to \code{"plate_id"}.}

\item{value_col}{Name of the numeric absorbance column. Defaults to
\code{"abs"}.}
}
\value{
A tibble with one row per plate, and columns \code{dataset}, \code{plate_id},
\code{blank_avg}, \code{blank_sdev}, \code{blank_coeff_var_percent}
}
\description{
This only makes sense for plates that have contain 2 or more standard curves for
the same standard solution
}
\examples{
tidy_plates
std_blank <- tidy_plates |> extract_std_blank(std_def = "Std")
std_blank$trusted |> std_blank_average()
# Don't be disturbed by the NA values: it is not possible to compute a
# standard deviation of coefficient of variation from only one data entry
}
