% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_corrected}
\alias{std_corrected}
\title{Example of Blank-corrected Data for Standard Curve}
\format{
A tibble with 1 row per well containing serial dilutions of the standard curve. Most relevant columns are described here

\describe{
\item{row, column, well_id, plate_id, dataset}{of the 96-well plate}
\item{unique_curve_id}{relevant when there are several curves per plate}
\item{unique_well_id}{combination of well_id and plate_id}
\item{map}{mapping og the wells ("Std" in this case)}
\item{abs_corrected}{blank-corrected absorbance value. Note that the \code{abs} column has been removed to avoid inadvertend use of uncorrected data }
\item{map}{mapping og the wells ("Std" in this case)}
\item{map}{mapping og the wells ("Std" in this case)}
\item{std_sp, std_unit, std_conc}{characterize the species, concentration
and unit of the standard solution in that well}
}
}
\usage{
std_corrected
}
\description{
Example of Blank-corrected Data for Standard Curve
}
\keyword{datasets}
