% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_std_curve.R
\name{suspicious_lm}
\alias{suspicious_lm}
\title{Extract Suspicious Rows from Linear Model Data (non-significance, Non-normality, heteroscedasticity)}
\usage{
suspicious_lm(lm_data, model = "linear")
}
\arguments{
\item{lm_data}{A tible containing data from the linear model. Structured as
the output of \code{lm_std_curve()}}

\item{model}{Which model to use. Accepts either \code{linear} (default) or \code{poly}
for polynomial model.}
}
\value{
A tible same structure as lm_data, but contains only "suspicious" standard curves
}
\description{
This "bad" subset of linear model data (regression of standard curves) is useful
in the case of numerous standard curves. It facilitates individual review
of suspicious curves only.
}
\examples{
data <- std_corrected |> dplyr::group_by(plate_id, column)
lm_data <- lm_std_curve(data)
suspicious_lm(lm_data)
}
\seealso{
\code{\link[=lm_std_curve]{lm_std_curve()}}

\code{\link[=plot_list_lm]{plot_list_lm()}}
}
