% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious_wells.R
\name{suspicious_wells}
\alias{suspicious_wells}
\title{Identifying wells for which raw absorbance is out of range}
\usage{
suspicious_wells(
  data,
  min_abs = 0.1,
  max_abs = 1,
  empty_wells = "empty",
  show_msg = TRUE,
  show_warning = TRUE,
  show_plot = TRUE,
  plot_binwidth = 0.01,
  plot_col_facet = "dataset",
  export_plot = NULL
)
}
\arguments{
\item{data}{Tibble, in a similar form to our \code{\link{tidy_table}}.
In particular, columns \code{map}, \code{abs}, \code{dataset}, \code{plate_id}, \code{well_id} must be present and named so.}

\item{min_abs, max_abs}{Lowest and highest accepted value for absorbance, default at 0.1 and 1, respectively.}

\item{empty_wells}{Character string corresponding to how empty wells are described in the plate mapping.
Defaults to "empty", can also be a vector containing several values. Note that wells described as \code{NA}
in the map column will also be ignored.}

\item{show_msg, show_warning}{Whether to keep (TRUE) or suppress (FALSE) the message or warning
that are returned in the case of absence (message) or presence (warning) of suspicious, out-of-range, wells}

\item{show_plot}{Whether to display the associated plot (histogram of absorbance values).
Default is \code{TRUE}. Note that the next arguments are only relevant is \code{show_plot = TRUE}.}

\item{plot_binwidth}{To set the binwidth of the histogram. Default is 0.01}

\item{plot_col_facet}{Which column to use to facet the plots. For no facetting,
choose \code{plot_col_facet = "none"} (For now, only facet with 1 axis is possible)}

\item{export_plot}{Defaults to null. Set it to a string to save the plot in the global environment,
the string will name the object (e.g., \code{export_plot = "abs_distrib"} will save an object called \code{abs_distrib})}
}
\value{
A table with 5 columns.
The first 3 (dataset, plate_id, well_id) allow the unique identification of suspicious wells
and can be used to run \code{\link[=remove_wells]{remove_wells()}}. The next 2 columns (map, abs) allow further visualization of those suspicious wells
}
\description{
\code{suspicious_wells()} extracts data relating to wells for which the absorbance is outside of a user-defined range.
Whereas there is a long-lasting tradition of setting "acceptable" absorbance values between 0.1 and 1 (default values for this function),
in reality, the acceptable range will depend on experiment, usage and quality of the spectrophotomer.
Note that values lower than 0.1 are not rare, especially for blank and lower values in the standard curve.
}
\examples{
# bring some NA (abs) and empty wells in tidy_table to check that those wells are removed
data <- tidy_table
data$abs[1] <- NA
data$map[2] <- "empty"
suspicious_wells(data)
}
