% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plates.R
\name{tecan_to_tibble}
\alias{tecan_to_tibble}
\title{Import plate data from a Tecan-generated .xlsx file}
\usage{
tecan_to_tibble(folderpath, extension = ".xlsx")
}
\arguments{
\item{folderpath}{Path to the folder that contains all .xlsx in the "Tecan"
output format. Make sure that folder contains only plate data as .xlsx files.
It may contain other file types, but all .xlsx files will be imported, which
could lead to errors should the files not present the correct structure}

\item{extension}{For now has only 1 .xlsx option. Should it be relevant to add
a csv option, please reach out to the authors.}
}
\value{
The plate data in a tibble format
}
\description{
Import plate data from a Tecan-generated .xlsx file
}
\examples{
filepath <- system.file("extdata", "tecan_example/", package = "plate2N")
tecan_to_tibble(filepath)

}
