% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tibble_example}
\alias{tibble_example}
\title{Example of a tibble format for plate data}
\format{
A tibble with 13 columns and 9 rows per plate, with no empty rows:
\describe{
\item{row}{column with an iterative structure every 9 rows. The first row of this column contains the plate id, the next 8 rows contain letters from A to H, reflecting plate rows}
\item{X1:X12}{containing data from the 12 columns of each plate, where the first row of each plate mentions plate-column number, from 1 to 12}
}
}
\usage{
tibble_example
}
\description{
This particular tibble was generated with the \code{csv_to_tibble()}function, but the end structure is strictly the same with all import options (e.g., with \code{skanit_to_tibble()}or \code{txt_to_tibble()}). It contains practise "plate data" to be used within the \code{plate2N}package. A tibble (data frame) with 96-well plate data in a plate format (data can be numerical, e.g., for absorbance data, or characters, e.g., for mapping data).
}
\keyword{datasets}
