% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_to_list.R
\name{tibble_to_list}
\alias{tibble_to_list}
\title{Converts plates data from tibble to a list}
\usage{
tibble_to_list(tibble)
}
\arguments{
\item{tibble}{The tibble containing all plate data. See output from examples under ?csv_import for a glimpse of the required tibble structure and column names}
}
\value{
A list where each element contains the data of a single plate, and the name of each element is the plate identifier (plate name). This list format is the same as the output from txt_import()
}
\description{
Converts a tibble (as outputted from csv_import() or skanit_to_plate()) containing all plate data into a list of tibbles (1 per plate)
}
\examples{
tibble <- csv_import(system.file("extdata", "csv_example.csv", package = "plate2N"))
tibble
list <- tibble_to_list(tibble)
names(list)[1]; list[[1]]
names(list)[2]; list[[2]]
}
