% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plates.R
\name{txt_to_tibble}
\alias{txt_to_tibble}
\title{Imports 96-well plate data from .TXT format as exported from the plate reader}
\usage{
txt_to_tibble(filepath, extension = ".TXT", output = "tibble")
}
\arguments{
\item{filepath}{The path to the folder containing the .TXT files. The folder
may contain other non-TXT files, but all .TXT files within the folder will be included.}

\item{extension}{The default is ".TXT". This parameter defines the pattern by which files to be included will be picked from the folder identified by 'filepath'.}

\item{output}{Desired output format. Default "tibble." Alternative option is a list (one element of the list attributed to each plate)}
}
\value{
Depends on the output parameter.
If \code{output = "tibble"}, a tibble where each plate is stacked on top of each other (same format as with skanit_to_plate())
If \code{output = "list"}, a list where each element is a tibble corresponding to raw plate data (1 file --> 1 plate --> 1 element). Names of the elements correspond to the names of the files (without the extension .TXT). A good practice is thus to name the files as "plate_name.TXT"
}
\description{
Imports 96-well plate data from .TXT format as exported from the plate reader
}
\examples{
filepath <- system.file("extdata", "txt_examples/", package = "plate2N")
abs_tibble <- txt_to_tibble(filepath)
}
