% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vertical_plates}
\alias{vertical_plates}
\title{Example of the tidy, verticalized plate data}
\format{
A tibble with 96 rows (1 per well of the 96-well plate) and 6 columns
\describe{
\item{row}{1-letter identifier of the original plate-row (letters A to H)}
\item{column}{1-number identifier of the original plate-column (numbers 1 to 12)}
\item{NO3_TDN_01 to NO3_TDN_04}{each column corresponds to 1 plate of the input tibble and contains verticalized plate data}
}
}
\usage{
vertical_plates
}
\description{
This data set was generated by the \code{verticalize_plates()}function, ran on our example data set \code{tibble_example}.
}
\keyword{datasets}
