% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verticalize_tidy_join.R
\name{verticalize_plates}
\alias{verticalize_plates}
\title{Tidying plate data (verticalization)}
\usage{
verticalize_plates(
  tibble,
  coerce_numeric = FALSE,
  prefix = NULL,
  row_col = "row",
  column_cols = paste0("X", 1:12)
)
}
\arguments{
\item{tibble}{The tibble containing the raw plate-formatted data to be tidied.
tibble must fit the following criteria:
- Plate ids \strong{cannot} be a single capital letter (e.g., "A", "B", ...)
- Plates must be complete (96 wells accounted for, set up in 12 columns x 8 rows), but may contain NA's
- Plate data in the tibble \strong{must be} structured exactly as in example files (see also \code{txt_to_tibble()}, \code{csv_to_tibble()} or \code{skanit_to_tibble()}).}

\item{coerce_numeric}{Whether or not to force data entries to be numerical. The default is set to \code{FALSE}, so that data will be outputted as strings}

\item{prefix}{Defaults as an empty string. A \code{prefix} can be added to all column names, which can be useful to join tables from distinct datasets}

\item{row_col}{Name of the column identifying plate row (A-H) in the
raw input. Defaults to \code{"row"}.}

\item{column_cols}{A vector of the 12 column names holding plate
columns 1-12 in the raw input, in order. Defaults to
\code{paste0("X", 1:12)}.}
}
\value{
A tidy tibble (verticalized plate data), with 1 column per plate
}
\description{
\code{verticalize_plates} brings plate data (absorbance or mapping data) into a vertical, tidy format.
It starts from a tibble in plate format (as rendered by \code{txt_to_tibble()}, \code{csv_to_tibble()} and \code{skanit_to_tibble()}),
and ends with a tidy tibble of 96 rows (row per well) and one column per plate, as well as structural columns allowing well identification (row, column)
}
\examples{
# check out input
tibble_example
(verticalize_plates(tibble_example))
(verticalize_plates(tibble_example, coerce_numeric = TRUE))
(verticalize_plates(tibble_example, prefix = "prefix_"))
}
\seealso{
\code{\link[=skanit_to_tibble]{skanit_to_tibble()}}, \code{\link[=csv_to_tibble]{csv_to_tibble()}}, \code{\link[=txt_to_tibble]{txt_to_tibble()}} that can generate the input needed
}
