% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_correct_abs.R
\name{blank_correct_abs}
\alias{blank_correct_abs}
\title{Transform Raw Absorbance Data into Blank-corrected Absorbance}
\usage{
blank_correct_abs(
  raw_wells_data,
  per_plate_avg_blank,
  map_to_exclude = c("empty", "Std", "extr")
)
}
\arguments{
\item{raw_wells_data}{A tibble, based on \code{tidy_plates}, can contain metadata as well.
\code{raw_wells_data} contains the raw absorbance data to be blank-corrected.
Must contain columns "abs" and "map".}

\item{per_plate_avg_blank}{Contains the per-plate average absorbance of the blank
Must contain column "blank_avg" (rename it prior to function call if needed)}

\item{map_to_exclude}{A vector of strings containing all \code{map} definitions of
wells that are not data per se (e.g., empty wells, etc.).
Defaults to \code{c("empty","Std","extr")}. If wells to exclude are not defined
by a unique "map" (e.g., blank wells of the standard curve), make sure to
filter out those rows from \code{raw_wells_data} before the function call.}
}
\value{
A tibble with the blank-corrected absorbance. It has the same structure
as \code{raw_wells_data}, but the \code{abs} column has been removed, and column
\code{abs_corrected} has been added. The output tibble normally contains less rows
than the input tibble (due to \code{map_to_exclude})
}
\description{
Transform Raw Absorbance Data into Blank-corrected Absorbance
}
\examples{
data <- tidy_plates
extractant_average <- tidy_plates |> extractant_average()
blank_correct_abs(
    raw_wells_data = data,
    per_plate_avg_blank = extractant_average |> dplyr::rename(blank_avg = extr_avg),
    map_to_exclude = c("empty","Std","extr"))
}
