% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw_extr.R
\name{boxplot_outlier_extr}
\alias{boxplot_outlier_extr}
\title{Identify outlier-wells within plates containing suspicious extractant data}
\usage{
boxplot_outlier_extr(suspicious_extractant, max_coeff = 5)
}
\arguments{
\item{suspicious_extractant}{A tibble as generated by \code{suspicious_extr()},
can be computed with \code{suspicious_extr()}}

\item{max_coeff}{User-defined threshold value, defaults at 5\%. All plates for which
the coefficient of variation for extractant raw absorbance is above this threshold
will be considered "suspicious plates". ! Be consistent with previous steps}
}
\value{
A boxplot generated by ggplot, highlighting information necessary to
create a tibble of wells to be removed with \code{remove_wells()}, i.e.,
for each boxplot: plate_ids, well_ids and plate number (corresponding to
the order of plate ids in \code{suspicious_extractant})
}
\description{
Identify outlier-wells within plates containing suspicious extractant data
}
\examples{
data <- tidy_plates
suspicious_plate_id <- qc_raw_extr(
   data, max_coeff = 5, suppress_message = TRUE, suppress_warning = TRUE)
suspicious_extr <- suspicious_extr(
   data, max_coeff = 5, suspicious_extr_per_plate = suspicious_plate_id)
boxplot_outlier_extr(
    suspicious_extractant = suspicious_extr,
    max_coeff = 5)

# case with 2 extractants
dbl_extr_plate
(suspicious_extr_per_plate <- qc_raw_extr(
    dbl_extr_plate, extr_def = c("extr_1", "extr_2"),
    max_coeff = 5, suppress_message = TRUE, suppress_warning = TRUE))
(suspicious_extr <- suspicious_extr(
    dbl_extr_plate, extr_def = c("extr_1", "extr_2"),
    max_coeff = 5, suspicious_extr_per_plate = suspicious_extr_per_plate))
boxplot_outlier_extr(
    suspicious_extractant = suspicious_extr,
    max_coeff = 5)
}
