% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_to_conc.R
\name{convert_molec}
\alias{convert_molec}
\title{Converts Concentrations From Compound Weight to Target Weight}
\usage{
convert_molec(conc_data, masses = molar_masses)
}
\arguments{
\item{conc_data}{A tibble containing columns "std_sp" (chr), "target_sp" (chr),
"conc_mgNsp_L" (dbl). Columns "target_sp" means "to convert to" (e.g., "N" for mg N / L),
and "std_sp" means "to convert from" (e.g., "NH4" for mg NH4 / L)}

\item{masses}{A named vector of molar masses (names of the elements = compounds.)}
}
\value{
A tibble of the same format, with an added column. For now, this new column
is named `conc_mgN_L". But we could consider adding an argument to define column names
}
\description{
Convert units of concentrations from "mg Compound L-1" to "mg Target L-1". This
allows, for example, to convert concentrations expressed in mg NH4 / L to
mg N / L. This conversion requires a vector where molar masses of both
Compound and Target are stored. This function works in bulk, on a tibble
Target (e.g., nitrogen) will be expressed in the same unit as the compound
(origin, e.g. NH4). For example, we go from "mg NH4 / L" to "mg N / L".
We don't go from mg to g or kg...
}
\examples{
tidy_plates
molar_masses
}
\seealso{
molar_masses
}
