% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_std_blank.R
\name{correct_std_blank}
\alias{correct_std_blank}
\title{Replaces raw absorbance data from standard curves by blank-corrected absorbance values.}
\usage{
correct_std_blank(
  data,
  std_def = "Std",
  pipetting_direction = "top_down",
  std_blanc = NULL
)
}
\arguments{
\item{data}{A tibble respecting the structure of \code{\link{tidy_table}}. \code{data} must have,
though not necessarily in that order, the following column names:
row, column, well_id, unique_well_id, dataset, plate_id, map, abs}

\item{std_def}{A string, defaults with \code{"Std"}: how data from wells containing the standard curve are referred to.}

\item{pipetting_direction}{Can only be "top_down" (default) or "bottom_up".
A top_down pipetting means that the curve was pipetted vertically (in a single column of the 96-well plate),
with the smallest value (blank) in row A and the highest value in row H.
Conversely, bottom_up pipetting would have the blank in row H and the most concentrated solution in row A}

\item{std_blanc}{If NULL (default), std_blanc will be computed within the function call.
Otherwise, \code{std_blanc} should be a tibble in the same format as \code{extract_std_blanc(data)$all}.
Changing the default value of \code{std_blanc} may be relevant if the previous
call to \code{\link[=extract_std_blanc]{extract_std_blanc()}} has led the user to correct "trusted" blancs
in any way (see \code{?extract_std_blanc()} for more details)}
}
\value{
A tibble with blank-corrected absorbance values for standard curves.
It has less rows than the input \code{data} because
- \code{correct_std_blank()} extracts standard curves-related data and
- for which it only keeps values for non-blank wells once the correction is done,
which is why row A (top_down pipetting) or row H (bottom_up pipetting) are missing from this output table
}
\description{
\code{correct_std_blank()} relies on \code{\link[=extract_std_blanc]{extract_std_blanc()}}.
}
\examples{
#tidy_plates
#correct_std_blank(tidy_plates)
}
\seealso{
\code{\link[=extract_std_blanc]{extract_std_blanc()}}
}
