% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_blanks.R
\name{extract_extractant}
\alias{extract_extractant}
\title{Extracting "extractant" data (blank for samples) from plate tidy table}
\usage{
extract_extractant(data, extr_def = "extr")
}
\arguments{
\item{data}{A tibble like \code{tidy_plates}}

\item{extr_def}{A string that characterizes wells containing the extractant
in the mapping (\code{map}column) of the plate}
}
\value{
A subset of \code{data}, containing only extractant data
}
\description{
Extracting "extractant" data (blank for samples) from plate tidy table
}
\examples{
data = tidy_plates
extract_extractant(data)
}
