% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_blanks.R
\name{extractant_average}
\alias{extractant_average}
\title{Computing the per-plate average for raw absorbance of the extractant (blank for samples)}
\usage{
extractant_average(data = NULL, extractant_data = NULL, extr_def = "extr")
}
\arguments{
\item{data}{A tibble like \code{tidy_plates}}

\item{extractant_data}{Alternatively, A tibble containing only extractant data
Defaults to NULL, where it would be computed from \code{data} using \code{extract_extractant(data)}}

\item{extr_def}{A string that characterizes wells containing the extractant
in the mapping (\code{map}column) of the plate}
}
\value{
A tibble with one row per plate, contianing the average, standard
deviation and coefficient of variation of raw absorbance of the extractant
}
\description{
Computing the per-plate average for raw absorbance of the extractant (blank for samples)
}
\examples{
data = tidy_plates
(extr_avg <- extractant_average(data))
}
