% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_std_curve.R
\name{lm_std_curve}
\alias{lm_std_curve}
\title{Perform Linear Model for Standard Curve}
\usage{
lm_std_curve(grouped_data, model = "linear")
}
\arguments{
\item{grouped_data}{A tibble, grouped per curve (e.g., use \code{dplyr::group_by(plate_id, column)}
on your data before calling the function). Must contain columns \code{std_conc},
\code{unique_curve_id} and \code{abs_corrected}}

\item{model}{Which model to use. Accepts either \code{linear} (default) or \code{poly}
for polynomial model.}
}
\value{
A table containing relevant parameters of the linear model, with
- 1 row per "group" (e.g., plate * column, which is relevant to spot outliers).
- columns: \code{unique_curve_id}, \code{slope}, \code{r_squared}, \code{adj_r_squared}, \code{lm_p},
\code{normality_lm_residuals}, \code{shapiro_p}, \code{homoscedasticity_lm_residuals},
\code{breusch_pagan_p}
}
\description{
The linear model is based on the function \code{lm()} and fits the curve to go through
the origin (0,0), which only makes sense for blank-corrected absorbance data.
}
\details{
Formula within \code{lm()}.
For linear model: \code{y = m*x + c};
For polynomial model: \code{y = a*x^2 + b*x + c};
With
- y = blank-corrected absorbance
- x = concentration
- c = 0 because we use blank-corrected absorbance data
}
\examples{
data <- std_corrected |> dplyr::group_by(plate_id, column)
lm_std_curve(data)
}
