% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw_extr.R
\name{multiplot_outlier_extr}
\alias{multiplot_outlier_extr}
\title{Multiple distribution plot to review suspicious extractant values and spot outliers}
\usage{
multiplot_outlier_extr(
  suspicious_extractant = NULL,
  data = NULL,
  max_coeff = 5
)
}
\arguments{
\item{suspicious_extractant}{A tibble as generated by \code{suspicious_extr()},
can be computed from \code{data} with \code{suspicious_extr()}}

\item{data}{Defaults to NULL. If provided, must be a tibble formatted as \code{tidy_plates}}

\item{max_coeff}{User-defined threshold value, defaults at 5\%. All plates for which
the coefficient of variation for extractant raw absorbance is above this threshold
will be considered "suspicious plates". ! Be consistent with previous steps}
}
\value{
A multiple plot for quality checking of extractant wells distribution
and definition of threshold values
}
\description{
Multiple distribution plot to review suspicious extractant values and spot outliers
}
\examples{
data <- tidy_plates
# 0.5 is unreasonable in most uses, but is used here to ensure some output
suspicious_extr_per_plate <- qc_raw_extr(data, max_coeff = 5,
    suppress_message = TRUE, suppress_warning = TRUE)
suspicious_extr <- suspicious_extr(data, max_coeff = 5,
    suspicious_extr_per_plate = suspicious_extr_per_plate)
multiplot_outlier_extr(suspicious_extractant = suspicious_extr, max_coeff = 5)
## Tip: if too many curves appear, consider cutting `suspicious_extractant`
##.     into several subsets, to be given as input for multiple runs of `multiplot_outlier_extr()`
}
