% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_std_curve.R
\name{plot_list_lm}
\alias{plot_list_lm}
\title{Compute a list of plots, one per standard curve}
\usage{
plot_list_lm(lm_data, std_data)
}
\arguments{
\item{lm_data}{lm_data in the shape as generated by \code{lm_std_curve()},
possibly as a subset to observe only suspicious curves}

\item{std_data}{tibble containing only standard data, as in \code{std_corrected}.
Absorbance values should already by blank-corrected}
}
\value{
A list of plots, one per curve, annotated with the issue:
- non-significance of the linear modal (p-value > 0.05)
- non-normality of residuals (shapiro test with threshold of p = 0.05)
- heteroscedasticity of residuals (Breusch-Pagan test, threshold of p = 0.05)
}
\description{
Compute a list of plots, one per standard curve
}
\examples{
lm_data <- lm_std_curve(std_corrected |> dplyr::group_by(plate_id, column))
plot_list <- plot_list_lm(
    lm_data,
    std_data = std_corrected)
plot_list[[1]] ; plot_list[[2]]
}
