% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve.R
\name{plot_std}
\alias{plot_std}
\title{Display one or more standard curve(s)}
\usage{
plot_std(std_data, through_origin = TRUE, model = "linear")
}
\arguments{
\item{std_data}{The table containing the data to be plotted. Must contain the
columns \code{std_conc}, \code{abs}, \code{plate_id} and \code{well_id}. If the plot shows too
many curves, consider filtering the input data frame or adding a ggplot layer
to facet (see \code{?facet_wrap()} or \code{?facet_grid()}).}

\item{through_origin}{Whether the smooth curve should be constrained to go
through the origin. Default to TRUE, which only makes sense for absorbance
data that has already been blank-corrected}

\item{model}{Which model to use for the smooth curve. Accepts either \code{linear}
(default) or \code{poly} for polynomial model (y = ax + bx^2 + c, with c = 0
if \code{through_origin = TRUE})}
}
\value{
A plot of one or several standard curves.
}
\description{
Plots raw absorbance (y-axis) vs concentration (x-axis), grouping the data by \code{plate_id}.
Uses the \code{ggplot2} package.
}
\examples{
raw_meta <- tidy_plates |>
    dplyr::left_join(metadata, by = dplyr::join_by(dataset,plate_id))
curve_concentration <- extract_curve(metadata)
std_data <- raw_meta |>
  extract_std_data() |>
  dplyr::select(!std_conc) |>
  dplyr::left_join(curve_concentration, by = dplyr::join_by(row, dataset, plate_id))
plot_std(std_data, through_origin = FALSE, model = "linear") + ggplot2::facet_wrap(dataset~plate_id)
}
