% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_to_conc.R
\name{reg_join_abs}
\alias{reg_join_abs}
\title{Merge Data from Regression Equation on Standard Curve with Absorbance Data (samples)}
\usage{
reg_join_abs(lm_table, abs_data, target_sp = "N")
}
\arguments{
\item{lm_table}{A tibble with linear model data (as generated by \code{lm_std_curve()}).
Must contain columns \code{plate_id}, \code{slope}, \code{adj_r_squared}, \code{lm_p}}

\item{abs_data}{Blank-corrected absorbance data, in a similar format to tidy_plates,
but containing only sample wells, no empty wells, standard curve wells or extractant wells}

\item{target_sp}{Defaults to "N", the target compound to convert to}
}
\value{
A tibble joining linear model data with absorbance data
}
\description{
Merge Data from Regression Equation on Standard Curve with Absorbance Data (samples)
}
\examples{
lm_table <- lm_output$lm_data
abs_data <- abs_corrected
reg_join_abs(lm_table, abs_data, target_sp = "N")
}
