% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve.R
\name{std_dilution_average}
\alias{std_dilution_average}
\title{Compute per-dilution Averages for Standard Curves}
\usage{
std_dilution_average(std_data)
}
\arguments{
\item{std_data}{A tibble of std data}
}
\value{
Same, with less rows (bc average of same-dilution wells per plate).
Artificial column 13
}
\description{
This needs better documentation!!
}
\examples{
std_corrected
std_dilution_average(std_corrected)
}
