% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw_extr.R
\name{suspicious_extr}
\alias{suspicious_extr}
\title{Extract suspicious extractant wells}
\usage{
suspicious_extr(data, suspicious_plate_id = NULL, max_coeff = 5)
}
\arguments{
\item{data}{Tibble formatted as \code{tidy_plates}}

\item{suspicious_plate_id}{If NULL (default), computed from \code{data} with \code{qc_raw_extr(data, max_coeff = max_coeff)}}

\item{max_coeff}{User-defined threshold value, defaults at 5\%. All plates for which
the coefficient of variation for extractant raw absorbance is above this threshold
will be considered "suspicious plates"}
}
\value{
A subset of \code{data} containing only plates where
raw extractant values should be reviewed
}
\description{
Extract suspicious extractant wells
}
\examples{
data <- tidy_plates
# 0.5 is unreasonable in most uses, but is used here to ensure some output
suspicious_plate_id <- qc_raw_extr(data, max_coeff = 0.5,
    suppress_message = TRUE, suppress_warning = TRUE)
suspicious_extr <- suspicious_extr(data, max_coeff = 0.5,
    suspicious_plate_id = suspicious_plate_id)
suspicious_extr
}
