% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_raw_extr.R
\name{suspicious_extr}
\alias{suspicious_extr}
\title{Review suspicious extractant values and spot outliers}
\usage{
suspicious_extr(data, suspicious_plate_id = NULL, max_coeff = 5)
}
\arguments{
\item{data}{Tibble formatted as \code{tidy_plates}}

\item{suspicious_plate_id}{If NULL (default), computed from \code{data} with \code{qc_raw_extr(data, max_coeff = max_coeff)}}

\item{max_coeff}{User-defined threshold value, defaults at 5\%. All plates for which
the coefficient of variation for extractant raw absorbance is above this threshold
will be considered "suspicious plates"}
}
\value{
A list with 2 elements.
\verb{$suspicious_extractant}, a subset of \code{data} containing only plates where
raw extractant values should be reviewed
\verb{$multiple_plot}, a collection of distributions of those raw absorbance values,
to help spotting outlier wells
}
\description{
Review suspicious extractant values and spot outliers
}
\examples{
data <- tidy_plates
# 0.5 is unreasonable in most uses, but is used here to ensure some output
suspicious_extr <- suspicious_extr(data, max_coeff = 0.5)
suspicious_extr$suspicious_extractant
suspicious_extr$multiple_plot

}
