% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verticalize_tidy_join.R
\name{vertical_to_tidy}
\alias{vertical_to_tidy}
\title{From vertical plate data to tidy data using the tidyr package}
\usage{
vertical_to_tidy(vertical_data)
}
\arguments{
\item{vertical_data}{As generated from either \code{\link[=verticalize_plates]{verticalize_plates()}} or \code{\link[=join_abs_map]{join_abs_map()}}.}
}
\value{
A table in a tidy format for downstream analysis
}
\description{
See also \code{\link{vertical_plates}} and \code{\link{tidy_table}} to understand input and output data structure
}
\examples{
map_file <- system.file("extdata", "csv_map.csv", package = "plate2N")
abs_folder <- system.file("extdata", "txt_examples/", package = "plate2N")
map_tibble <- csv_to_tibble(map_file)
abs_tibble <- txt_to_tibble(abs_folder)
joined_vertical <- join_abs_map(abs_tibble, map_tibble, dataset = "Nmin-")
tidy_data <- vertical_to_tidy(joined_vertical)
}
