% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verticalize_plates.R
\name{verticalize_plates}
\alias{verticalize_plates}
\title{Tidying plate data (verticalization)}
\usage{
verticalize_plates(tibble)
}
\arguments{
\item{tibble}{The tibble containing the raw plate-formatted data to be tidied.
tibble must fit the following criteria:
- Plate ids \strong{cannot} be a single capital letter (e.g., "A", "B", ...)
- Plates must be complete (96 wells accounted for, set up in 12 columns x 8 rows), but may contain NA's
- Plate data in the tibble \strong{must be} structured exactly as in example files (see also \code{txt_to_tibble()}, \code{csv_to_tibble()} or \code{skanit_to_tibble()}).}
}
\value{
A tidy tibble (verticalized plate data), with 1 column per plate
}
\description{
\code{verticalize_plates}brings plate data (absorbance or mapping data) into a vertical, tidy format.
It starts from a tibble in plate format (as rendered by \code{txt_to_tibble()}, \code{csv_to_tibble()} and \code{skanit_to_tibble()}),
and ends with a tidy tibble of 96 rows (row per well) and one column per plate, as well as structural columns allowing well identification (row, column)
}
\examples{
# create tibble with skanit_to_tibble()
skanit_csv <- system.file("extdata", "skanit.csv", package = "plate2N")
plate_data <- skanit_to_tibble(skanit_csv, delim = ",")
(tibble <- plate_data$abs_tibble)

# run verticalize_plates()
(verticalize_plates(tibble))

}
